chore: refacto tests
